Add fix for #2 - back and forward navigation not firing events
There appears to be a Polymer timing issue where observed properties in
this case (route) do not correctly trigger a routeChanged event were
expected.

In this case my workaround is to manually trigger the event fire when
we know for sure that the route has changed. A workaround is being
added for another issue this introduces around routes on page load.